requirements.txt
# Main requirements for training, inference, and Streamlit app
streamlit
tensorflow
Pillow
numpy
matplotlib
seaborn
scikit-learn
